requirements.txt
Flask==3.0.3
google-genai==1.7.0
gunicorn==22.0.0
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.17.2
google-cloud-language==2.15.1

